+++ test/controllers/admin_controller_test.rb
@@ -19,13 +19,18 @@ class AdminControllerTest < ActionDispatch::IntegrationTest
   test "should not get dashboard as editor" do
     login(@editor)
     get admin_url
-    assert_redirected_to root_path
+    assert_redirected_to dashboard_url
   end
 
   test "should not get dashboard as regular" do
     login(@regular)
     get admin_url
-    assert_redirected_to root_path
+    assert_redirected_to dashboard_url
+  end
+
+  test "should not get dashboard as public user" do
+    get admin_url
+    assert_redirected_to new_user_session_url
   end
 
   test "should get debug as admin" do