+++ test/test_helper.rb
@@ -55,3 +55,17 @@ class ActionDispatch::SystemTestCase
     click_form_submit
   end
 end
+
+# Discarding files stored during integration tests.
+module ActionDispatch
+  class IntegrationTest
+    def remove_uploaded_files
+      FileUtils.rm_rf(Rails.root.join("tmp", "storage"))
+    end
+
+    def after_teardown
+      super
+      remove_uploaded_files
+    end
+  end
+end